+++ test/Makefile
@@ -1,7 +1,7 @@
 all: basiccreate basicread basicalloc
 
 clean:
-	rm -f basicreate basicread basicalloc *.o *.lst
+	rm -f basicreate basicread basicalloc *.o *.lst *.dbg
 
 INCDIR=../include
 
@@ -14,12 +14,12 @@ basiccreate: basiccreate.o ../src/*.o
 basicalloc: basicalloc.o ../src/*.o
 	$(LD) $(LDFLAGS) -o $@ $< ../src/*.o
 
-basicread.o: basicread.c $(INCDIR)/ioservices.h $(INCDIR)/*.h
+basicread.o: basicread.c $(INCDIR)/iosvcs.h $(INCDIR)/*.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -c $<
 
-basiccreate.o: basiccreate.c $(INCDIR)/ioservices.h $(INCDIR)/*.h
+basiccreate.o: basiccreate.c $(INCDIR)/iosvcs.h $(INCDIR)/*.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -c $<
 
-basicalloc.o: basicalloc.c $(INCDIR)/ioservices.h $(INCDIR)/*.h
+basicalloc.o: basicalloc.c $(INCDIR)/iosvcs.h $(INCDIR)/*.h
 	$(CC) $(CFLAGS) $(CPPFLAGS) -I$(INCDIR) -c $<
